backend/requirements.txt
fastapi==0.115.0
mangum==0.19.0
pydantic==2.9.0
boto3==1.35.0
curl_cffi==0.7.1
beautifulsoup4==4.12.3
lxml==5.3.0
uvicorn==0.30.0
python-dotenv==1.0.1
